Update qubes-doc-folders.txt
Suggestion for a folder, qubes-projects, which allows sub-folders with more individual freedom on content-structure, for as long it is kept inside this single sub-folder for every project. It also makes it easier to find all Qubes projects in a single location, whether briefly documented, or extensively documented, keeping them all together makes it easier to navigate. For example Qubes-TV may have a couple of different documents here, or Qubes-Binary-Router documentations. Projects like these and other projects that people may put up, or found elsewhere on the internet, could fit neatly in such a specialized folder. Thoughts on this? 

Examples
- /contents/docs/qubes-projects/qubes-tv/broader-content-structure-freedom.
- /contents/docs/qubes-projects/qubes-binary-router/broader-content-structure-freedom.
- etc.