vpn.md: Fix hang on proxyVM shutdown
I changed qubes-vpn-handler.sh to check for a scheduled shutdown.
This will cause a clean exit of the openvpn down option in case the VM is requested to shut down.
I use a debian-11-minimal-template on Qubes 4.1 for the openvpn-vm.
Without my workaround the shutdown (e.g. via qvm-shutdown) hangs for the default 90 seconds with "A stop job is running for Qubes mis...post-boot actions".