Update arch building guide to be more concise
The archlinux building guide had a lot of different format types and
styles of writing which make it harder to read than necessary.
Furhtermore, some information was added that is of no special interest
there.

This commit tries to use a consistent and simple style for formatting
the content and uses wordings that tells the reader what to do.